user register form and view

+++ src/comments/views.py
@@ -1,9 +1,9 @@
 from django.contrib import messages
+from django.core.exceptions import ObjectDoesNotExist
 
 from trydjango19.helpers import get_or_create_new_comment
 
-from django.contrib.contenttypes.models import ContentType
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404, HttpResponse
 from django.shortcuts import render, get_object_or_404
 
 from .forms import CommentForm
@@ -11,7 +11,14 @@ from .models import Comment
 
 
 def comment_thread(request, id):
-    obj = get_object_or_404(Comment, id=id)
+    try:
+        obj = Comment.objects.get(id=id)
+    except ObjectDoesNotExist:
+        raise Http404
+
+    if not obj.is_parent:
+        obj = obj.parent
+
     initial_data = {"content_type": obj.content_type, "object_id": obj.object_id}
     form = CommentForm(request.POST or None, initial=initial_data)
     if form.is_valid():
@@ -23,7 +30,14 @@ def comment_thread(request, id):
 
 
 def comment_delete(request, id):
-    obj = get_object_or_404(Comment, id=id)
+    try:
+        obj = Comment.objects.get(id=id)
+    except ObjectDoesNotExist:
+        raise Http404
+
+    if obj.user != request.user:
+        return HttpResponse(status=403, content="you do not have permission")
+
     if request.method == "POST":
         parent_obj_url = obj.content_object.get_absolute_url()
         obj.delete()
